[REFACTOR] Move to workflows package
